Changes factory API
