+++ src/bundle-with-react.js
@@ -1,10 +1,10 @@
 import partial from 'lodash.partial';
 import mapValues from 'lodash.mapvalues';
 
-import { createElement } from 'react';
+import reactOptions from './util/react';
 import * as components from './';
 
 export const {
   createSlide,
   createTitle
-} = mapValues(components, (component) => partial(component, createElement));
+} = mapValues(components, (component) => partial(component, reactOptions));
