+++ src/util/react.js
@@ -1,0 +1,7 @@
+import React, { createElement } from 'react';
+
+export default {
+  h: createElement,
+  map: (children, fun) => React.Children.map(React.Children.only(children), fun)[0],
+  clone: React.cloneElement
+};
